requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (Updated for Python 3.13 compatibility)
sqlalchemy==2.0.35
alembic==1.13.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email validation
pydantic[email]==2.6.0

# PDF Processing (based on 2025 research - pypdf is now maintained)
pypdf==4.3.1
python-docx==1.1.2

# AI/Claude API
anthropic==0.40.0

# Environment and config
python-dotenv==1.0.1
pydantic-settings==2.6.0

# Templates
jinja2==3.1.4